requirements.txt
python-dotenv
openai
gradio
huggingface_hub<0.26
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-community
datasets==3.6.0
google-generativeai
anthropic